# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/reels": {
      "get": {
        "description": "feed=following = เฉพาะคนที่ติดตาม · author_username = หน้าโปรไฟล์ของคนนั้น",
        "operationId": "ReelsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 2,
              "type": "number"
            }
          },
          {
            "name": "per_page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 50,
              "type": "number"
            }
          },
          {
            "name": "feed",
            "required": false,
            "in": "query",
            "description": "following = เฉพาะคลิปของคนที่ฉันติดตาม (รวมของตัวเอง) — ฟีดแบบ Instagram",
            "schema": {
              "default": "all",
              "type": "string",
              "enum": [
                "all",
                "following"
              ]
            }
          },
          {
            "name": "author_username",
            "required": false,
            "in": "query",
            "description": "เฉพาะคลิปของคนนี้ — ใช้ทำหน้าโปรไฟล์",
            "schema": {
              "example": "6704101382-anuchat",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "สำเร็จ พร้อมข้อมูลแบ่งหน้า",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ReelResponseDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMeta"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "ฟีดคลิปสั้น เรียงใหม่ไปเก่า",
        "tags": [
          "reels"
        ]
      },
      "post": {
        "operationId": "ReelsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReelDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "สร้างคลิปสำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/ReelResponseDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "ไฟล์ยังไม่พร้อม หรือไม่ใช่วิดีโอ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "not_found"
                        },
                        "message": {
                          "type": "string",
                          "example": "ไม่พบคลิปนี้ อาจถูกลบไปแล้ว"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "ไฟล์นี้ไม่ใช่ของผู้เรียก",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "not_found"
                        },
                        "message": {
                          "type": "string",
                          "example": "ไม่พบคลิปนี้ อาจถูกลบไปแล้ว"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "โพสต์คลิปใหม่จากไฟล์ที่อัปโหลดเสร็จแล้ว",
        "tags": [
          "reels"
        ]
      }
    },
    "/api/v1/reels/{id}": {
      "get": {
        "operationId": "ReelsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/ReelResponseDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "ไม่พบคลิป",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "not_found"
                        },
                        "message": {
                          "type": "string",
                          "example": "ไม่พบคลิปนี้ อาจถูกลบไปแล้ว"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "รายละเอียดคลิปเดียว",
        "tags": [
          "reels"
        ]
      },
      "delete": {
        "operationId": "ReelsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "ลบคลิปของคนอื่นไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "not_found"
                        },
                        "message": {
                          "type": "string",
                          "example": "ไม่พบคลิปนี้ อาจถูกลบไปแล้ว"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "ลบคลิป (เจ้าของ หรือ admin ระดับองค์กร)",
        "tags": [
          "reels"
        ]
      }
    },
    "/api/v1/reels/{id}/likes": {
      "post": {
        "operationId": "ReelsController_like",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/LikeCountDto"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "กดไลก์คลิป (กดซ้ำไม่เพิ่มยอด)",
        "tags": [
          "reels"
        ]
      },
      "delete": {
        "operationId": "ReelsController_unlike",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/LikeCountDto"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "เลิกไลก์คลิป",
        "tags": [
          "reels"
        ]
      }
    },
    "/api/v1/reels/{id}/views": {
      "post": {
        "description": "นับคนที่เคยดู ไม่ใช่จำนวนครั้งที่เล่น — คนเดิมเรียกซ้ำยอดไม่ขยับ จึงปั่นด้วยการรีเฟรชไม่ได้",
        "operationId": "ReelsController_markViewed",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "บันทึกแล้ว",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/ViewCountDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "ไม่พบคลิปนี้",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "not_found"
                        },
                        "message": {
                          "type": "string",
                          "example": "ไม่พบคลิปนี้ อาจถูกลบไปแล้ว"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "บันทึกว่าดูคลิปแล้ว",
        "tags": [
          "reels"
        ]
      }
    },
    "/api/v1/reels/{id}/comments": {
      "get": {
        "operationId": "ReelsController_listComments",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 2,
              "type": "number"
            }
          },
          {
            "name": "per_page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "สำเร็จ พร้อมข้อมูลแบ่งหน้า",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ReelCommentResponseDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMeta"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "ความคิดเห็นใต้คลิป (ใหม่สุดก่อน)",
        "tags": [
          "reels"
        ]
      },
      "post": {
        "operationId": "ReelsController_addComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReelCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/ReelCommentResponseDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "ไม่พบคลิปนี้",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "not_found"
                        },
                        "message": {
                          "type": "string",
                          "example": "ไม่พบคลิปนี้ อาจถูกลบไปแล้ว"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "คอมเมนต์ใต้คลิป",
        "tags": [
          "reels"
        ]
      }
    },
    "/api/v1/reels/{id}/comments/{commentId}": {
      "delete": {
        "operationId": "ReelsController_removeComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "ลบความคิดเห็นของคนอื่นไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "not_found"
                        },
                        "message": {
                          "type": "string",
                          "example": "ไม่พบคลิปนี้ อาจถูกลบไปแล้ว"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "ลบความคิดเห็น (เจ้าของ หรือผู้ดูแล)",
        "tags": [
          "reels"
        ]
      }
    },
...
  },
  "info": {
    "title": "CS Nexus — aie4-social-reels",
    "description": "ระบบย่อยภายใต้ CSMJU2030 · ตัวตนมาจาก API Gateway ผ่าน header X-User-Id / X-Layer1-Role / X-Faculty",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "current_page": {
            "type": "number",
            "example": 2
          },
          "per_page": {
            "type": "number",
            "example": 50
          },
          "total_pages": {
            "type": "number",
            "example": 50
          },
          "total_items": {
            "type": "number",
            "example": 2480
          }
        },
        "required": [
          "current_page",
          "per_page",
          "total_pages",
          "total_items"
        ]
      },
...
      "ReelResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "caption": {
            "type": "object",
            "nullable": true
          },
          "asset_id": {
            "type": "string"
          },
          "duration_ms": {
            "type": "number"
          },
          "author_username": {
            "type": "string",
            "description": "Shared Identity ของเจ้าของคลิป — ไม่มี student_id ซ้ำซ้อน",
            "example": "6704101382-anuchat"
          },
          "like_count": {
            "type": "number"
          },
          "view_count": {
            "type": "number"
          },
          "liked_by_me": {
            "type": "boolean",
            "description": "ผู้เรียกกดไลก์คลิปนี้ไว้หรือยัง"
          },
          "created_at": {
            "type": "string",
            "description": "ISO 8601 พร้อม offset ตามมาตรฐาน",
            "example": "2026-08-11T09:30:00+07:00"
          }
        },
        "required": [
          "id",
          "title",
          "caption",
          "asset_id",
          "duration_ms",
          "author_username",
          "like_count",
          "view_count",
          "liked_by_me",
          "created_at"
        ]
      },
      "CreateReelDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "โชว์ UI ฟีด Reels ด้วย Tailwind"
          },
          "caption": {
            "type": "string",
            "example": "ทำด้วย Next.js + Framer Motion #CSNexus"
          },
          "asset_id": {
            "type": "string",
            "description": "id ของ Asset ที่ commit แล้วและสถานะเป็น READY",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "duration_ms": {
            "type": "number",
            "description": "ความยาวคลิปเป็นมิลลิวินาที สูงสุด 60 วินาที",
            "example": 28500
          }
        },
        "required": [
          "title",
          "asset_id",
          "duration_ms"
        ]
      },
      "LikeCountDto": {
        "type": "object",
        "properties": {
          "like_count": {
            "type": "number",
            "example": 129
          }
        },
        "required": [
          "like_count"
        ]
      },
      "ViewCountDto": {
        "type": "object",
        "properties": {
          "view_count": {
            "type": "number",
            "example": 842,
            "description": "จำนวนคนที่เคยดู ไม่ใช่จำนวนครั้งที่เล่น — คนหนึ่งนับครั้งเดียว"
          }
        },
        "required": [
          "view_count"
        ]
      },
      "ReelCommentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "reel_id": {
            "type": "string"
          },
          "author_username": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "reel_id",
          "author_username",
          "content",
          "created_at"
        ]
      },
      "CreateReelCommentDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "example": "ตัดต่อดีมากครับ"
          }
        },
        "required": [
          "content"
        ]
      },
...
    }
  }
}